remove theme props in editor component, deprecated lightColor and darkColor
